refactor: migrate constants to env configuration and clean up imports
