+++ src/deploy-commands.ts
@@ -1,11 +1,11 @@
 import { REST, Routes } from "discord.js";
-import { config } from "./config";
+import { env } from "@/env";
 import { commands } from "./commands";
 import logger from "./lib/logger";
 
 const commandsData = Object.values(commands).map((command) => command.data);
 
-const rest = new REST({ version: "10" }).setToken(config.DISCORD_TOKEN);
+const rest = new REST({ version: "10" }).setToken(env.DISCORD_TOKEN);
 
 type DeployCommandsProps = {
   guildId: string;
@@ -16,7 +16,7 @@ export async function deployCommands({ guildId }: DeployCommandsProps) {
     logger.info("Started refreshing application (/) commands.");
 
     await rest.put(
-      Routes.applicationGuildCommands(config.DISCORD_CLIENT_ID, guildId),
+      Routes.applicationGuildCommands(env.DISCORD_CLIENT_ID, guildId),
       {
         body: commandsData,
       }
